# swagger/jmarkets.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "JMarkets API Documentation",
    "version": "1.0.0",
    "description": "Multi-tenant marketplace platform API with CMS, authentication, RBAC, and auto-deployment",
    "contact": {
      "name": "JMarkets Support",
      "url": "https://j-markets.jcampos.dev"
    }
  },
  "servers": [
    {
      "url": "https://markets-api.jcampos.dev",
      "description": "Production API"
    },
    {
      "url": "http://localhost:5000",
      "description": "Development API"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
...
  },
  "paths": {
...
    "/api/users/{userId}/organization/{orgId}/pages/{pageId}/sections/{sectionId}/content": {
      "get": {
        "summary": "Get all content for a section",
        "tags": [
          "Section Content"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "orgId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "pageId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "sectionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of content items"
          },
          "404": {
            "description": "Section not found"
          }
        }
      }
    },
    "/api/users/{userId}/organization/{orgId}/pages/{pageId}/sections/{sectionId}/content/bulk": {
      "post": {
        "summary": "Bulk upsert content for a section",
        "tags": [
          "Section Content"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "orgId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "pageId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "sectionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "content"
                ],
                "properties": {
                  "content": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "key",
                        "value",
                        "displayName",
                        "valueType"
                      ],
                      "properties": {
                        "key": {
                          "type": "string"
                        },
                        "value": {
                          "type": "string"
                        },
                        "displayName": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        },
                        "valueType": {
                          "type": "string",
                          "enum": [
                            "text",
                            "color",
                            "image",
                            "boolean",
                            "json",
                            "background"
                          ]
                        },
                        "componentId": {
                          "type": "string"
                        },
                        "sortOrder": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Content upserted successfully"
          },
          "400": {
            "description": "Invalid data"
          },
          "404": {
            "description": "Section not found"
          }
        }
      }
    },
    "/api/users/{userId}/organization/{orgId}/pages/{pageId}/sections/{sectionId}/content/{contentId}": {
      "put": {
        "summary": "Update a content item",
        "tags": [
          "Section Content"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "orgId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "pageId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "sectionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "contentId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "key": {
                    "type": "string"
                  },
                  "value": {
                    "type": "string"
                  },
                  "displayName": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "valueType": {
                    "type": "string"
                  },
                  "componentId": {
                    "type": "string"
                  },
                  "sortOrder": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Content updated successfully"
          },
          "404": {
            "description": "Content not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a content item",
        "tags": [
          "Section Content"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "orgId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "pageId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "sectionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "contentId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Content deleted successfully"
          },
          "404": {
            "description": "Content not found"
          }
        }
      }
    },
    "/api/users/{userId}/organization/{orgId}/pages/{pageId}/sections": {
      "get": {
        "summary": "Get all sections for a page",
        "tags": [
          "Sections"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "orgId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "pageId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "activeOnly",
            "schema": {
              "type": "boolean"
            },
            "description": "Return only active sections"
          }
        ],
        "responses": {
          "200": {
            "description": "List of sections"
          },
          "404": {
            "description": "Page not found"
          }
        }
      },
      "post": {
        "summary": "Create a new section",
        "tags": [
          "Sections"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "orgId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "pageId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "sectionType",
                  "name"
                ],
                "properties": {
                  "sectionType": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "sortOrder": {
                    "type": "integer"
                  },
                  "isActive": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Section created successfully"
          },
          "400": {
            "description": "Invalid data"
          },
          "404": {
            "description": "Page not found"
          }
        }
      }
    },
    "/api/users/{userId}/organization/{orgId}/pages/{pageId}/sections/{sectionId}": {
      "put": {
        "summary": "Update a section",
        "tags": [
          "Sections"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "orgId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "pageId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "sectionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "sectionType": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "sortOrder": {
                    "type": "integer"
                  },
                  "isActive": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Section updated successfully"
          },
          "404": {
            "description": "Section not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a section",
        "tags": [
          "Sections"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "orgId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "pageId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "sectionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Section deleted successfully"
          },
          "404": {
            "description": "Section not found"
          }
        }
      }
    },
...
  },
...
}